tetragon: Move d_path into bpf_d_path.h header
Moving d_path into bpf_d_path.h header so it can be easily used in
following changes from other places.
