+++ bpf/lib/bpf_task.h
@@ -9,19 +9,6 @@
 #include "generic.h"
 #include "vmlinux.h"
 
-/* __d_path_local flags */
-// #define UNRESOLVED_MOUNT_POINTS	   0x01 // (deprecated)
-// this error is returned by __d_path_local in the following cases:
-// - the path walk did not conclude (too many dentry)
-// - the path was too long to fit in the buffer
-#define UNRESOLVED_PATH_COMPONENTS 0x02
-
-#ifdef __LARGE_BPF_PROG
-#define PROBE_CWD_READ_ITERATIONS 128
-#else
-#define PROBE_CWD_READ_ITERATIONS 11
-#endif
-
 FUNC_INLINE struct task_struct *get_parent(struct task_struct *t)
 {
 	struct task_struct *task;
